refactor(bot): split handlers into modules
